Implement Noth's recommendations
Move MonsterTask checking/starting to from send_map_get_info to Map where it belongs.
Moved MapPosition class to it's own file.
Changed return to log an error and the continue.
